moving _ir.json from mediafolder to addonfolder as config.json
If there is no config.json in addonfolder it will be tried to readout _ir.json from mediafolder. after that it will be deleted and the settings will be stored to config.json on next closing of anki.
If there is no config.json and no _ir.json it self.defaults will be loaded and saved to to config.json on next closing of anki.
To catch up an exception default settings will also be loaded if the json file is corrupt somehow.

This is just to get the plugin to meet new standards of the Anki 2.1 addons. It is not important, but was one of the things I played around.